requirements.txt
streamlit
python-dotenv
elasticsearch

